count_meter: Implemented strib's PR feedback.
